Refactor data-retrieval
